[Frontend] component 렌더링 방식 수정 (#198)
* fix(App): reset on 401

* fix: change structure by changed render method

* fix: change structure by changed render method

+++ frontend/src/App.js
@@ -33,9 +33,11 @@ export default class App extends Component {
           }
         })
         .catch((e) => {
-          sessionStorage.clear();
-          navigate("/start");
-          showToast(e);
+          if (e.status && e.status === 401) {
+            sessionStorage.clear();
+            navigate("/start");
+            showToast(e);
+          }
         });
     };
